fix: replaced uint with uint64 for improved compatibility (#454)

+++ src/qldpc/codes/distance_test.py
@@ -29,8 +29,8 @@ import qldpc
 
 
 def _bitwise_count(
-    val: npt.ArrayLike, out: npt.NDArray[np.uint] | None = None
-) -> npt.NDArray[np.uint]:
+    val: npt.ArrayLike, out: npt.NDArray[np.uint64] | None = None
+) -> npt.NDArray[np.uint64]:
     """Simplistic implementation of `bitwise_count` used to validate optimized variants."""
     val = np.asarray(val)
     nbits = 8 * val.itemsize
@@ -46,7 +46,7 @@ def _bitwise_count(
 
 
 def test_hamming_weight() -> None:
-    vals = np.random.randint(0, 2**64, size=(7, 11), dtype=np.uint)
+    vals = np.random.randint(0, 2**64, size=(7, 11), dtype=np.uint64)
     expected_weights = _bitwise_count(vals)
 
     weights = qldpc.codes.distance._hamming_weight(vals)
@@ -63,9 +63,9 @@ def test_hamming_weight() -> None:
 
 
 def test_symplectic_weight() -> None:
-    vals = np.random.randint(0, 2**64, size=(7, 11), dtype=np.uint)
+    vals = np.random.randint(0, 2**64, size=(7, 11), dtype=np.uint64)
     weights = qldpc.codes.distance._symplectic_weight(vals)
-    expected_weights = _bitwise_count((vals | (vals >> np.uint(1))) & 0x5555555555555555)
+    expected_weights = _bitwise_count((vals | (vals >> np.uint64(1))) & 0x5555555555555555)
     np.testing.assert_array_equal(weights, expected_weights)
 
     weight_fn = np.vectorize(qldpc.codes.distance._symplectic_weight_single, signature="()->()")
@@ -79,7 +79,7 @@ def test_symplectic_weight() -> None:
 
 
 def test_get_hamming_weight_fn() -> None:
-    generators = np.random.randint(2, size=(4, 64), dtype=np.uint)
+    generators = np.random.randint(2, size=(4, 64), dtype=np.uint64)
     weight_fn, nbuf = qldpc.codes.distance._get_hamming_weight_fn()
     weights_default = weight_fn(generators)
 
@@ -123,7 +123,7 @@ def test_get_hamming_weight_fn() -> None:
 
 
 def test_get_symplectic_weight_fn() -> None:
-    generators = np.random.randint(2, size=(4, 56), dtype=np.uint)
+    generators = np.random.randint(2, size=(4, 56), dtype=np.uint64)
     weight_fn, nbuf = qldpc.codes.distance._get_symplectic_weight_fn()
     weights_default = weight_fn(generators)
 
@@ -176,7 +176,7 @@ def test_get_symplectic_weight_fn() -> None:
 
 @pytest.mark.parametrize(
     "base_val",
-    [1, 2**64 - 1, int(np.random.randint(2**64, dtype=np.uint)) | 1],
+    [1, 2**64 - 1, int(np.random.randint(2**64, dtype=np.uint64)) | 1],
 )
 def test_count_trailing_zeros(base_val: int) -> None:
     for i in range(128):
@@ -185,7 +185,7 @@ def test_count_trailing_zeros(base_val: int) -> None:
 
 @pytest.mark.parametrize("width", range(1, 8))
 def test_inplace_rowsum(width: int) -> None:
-    arr = np.random.randint(2**31, size=(10, width), dtype=np.uint)
+    arr = np.random.randint(2**31, size=(10, width), dtype=np.uint64)
     expected = arr.sum(-1)
     actual = qldpc.codes.distance._inplace_rowsum(arr)
     np.testing.assert_array_equal(actual, expected)
@@ -202,7 +202,7 @@ def test_rows_to_ints_endianness() -> None:
     np.testing.assert_array_equal(ints, expected)
 
 
-@pytest.mark.parametrize("dtype", [int, np.uint, np.uint8, np.int16])
+@pytest.mark.parametrize("dtype", [int, np.uint64, np.uint8, np.int16])
 def test_rows_to_ints(dtype: npt.DTypeLike) -> None:
     bits = np.random.randint(2, size=(10, 93))
     ints = qldpc.codes.distance._rows_to_ints(bits, dtype=dtype)
@@ -250,10 +250,10 @@ def test_get_distance_classical(block_size: int) -> None:
     observed_bitstrings: list[tuple[int, ...]] = []
 
     def _mock_hamming_weight(
-        arr: npt.NDArray[np.uint],
-        buf: npt.NDArray[np.uint] | None = None,
-        out: npt.NDArray[np.uint] | None = None,
-    ) -> npt.NDArray[np.uint]:
+        arr: npt.NDArray[np.uint64],
+        buf: npt.NDArray[np.uint64] | None = None,
+        out: npt.NDArray[np.uint64] | None = None,
+    ) -> npt.NDArray[np.uint64]:
         observed_bitstrings.extend(map(tuple, arr.tolist()))
         return qldpc.codes.distance._hamming_weight(arr, buf=buf, out=out)
 
@@ -272,7 +272,7 @@ def test_get_distance_classical(block_size: int) -> None:
     assert len(observed_bitstrings) == len(expected_bitstrings)
     assert set(observed_bitstrings) == set(expected_bitstrings)
 
-    observed_array = np.array(observed_bitstrings, dtype=np.uint)
+    observed_array = np.array(observed_bitstrings, dtype=np.uint64)
     expected_distance = _bitwise_count(observed_array).sum(-1).min()
     assert distance == expected_distance
 
@@ -287,10 +287,10 @@ def test_get_distance_quantum(block_size: int) -> None:
     observed_bitstrings: list[tuple[int, ...]] = []
 
     def _mock_hamming_weight(
-        arr: npt.NDArray[np.uint],
-        buf: npt.NDArray[np.uint] | None = None,
-        out: npt.NDArray[np.uint] | None = None,
-    ) -> npt.NDArray[np.uint]:
+        arr: npt.NDArray[np.uint64],
+        buf: npt.NDArray[np.uint64] | None = None,
+        out: npt.NDArray[np.uint64] | None = None,
+    ) -> npt.NDArray[np.uint64]:
         observed_bitstrings.extend(map(tuple, arr.tolist()))
         return qldpc.codes.distance._hamming_weight(arr, buf=buf, out=out)
 
@@ -314,7 +314,7 @@ def test_get_distance_quantum(block_size: int) -> None:
     assert len(observed_bitstrings) == len(expected_bitstrings)
     assert set(observed_bitstrings) == set(expected_bitstrings)
 
-    observed_array = np.array(observed_bitstrings, dtype=np.uint)
+    observed_array = np.array(observed_bitstrings, dtype=np.uint64)
     expected_distance = _bitwise_count(observed_array).sum(-1).min()
     assert distance == expected_distance
 
@@ -329,10 +329,10 @@ def test_get_distance_quantum_symplectic(block_size: int) -> None:
     observed_bitstrings: list[tuple[int, ...]] = []
 
     def _mock_symplectic_weight(
-        arr: npt.NDArray[np.uint],
-        buf: npt.NDArray[np.uint] | None = None,
-        out: npt.NDArray[np.uint] | None = None,
-    ) -> npt.NDArray[np.uint]:
+        arr: npt.NDArray[np.uint64],
+        buf: npt.NDArray[np.uint64] | None = None,
+        out: npt.NDArray[np.uint64] | None = None,
+    ) -> npt.NDArray[np.uint64]:
         observed_bitstrings.extend(map(tuple, arr.tolist()))
         return qldpc.codes.distance._symplectic_weight(arr, buf=buf, out=out)
 
@@ -364,7 +364,7 @@ def test_get_distance_quantum_symplectic(block_size: int) -> None:
 
 
 def test_get_distance_classical_methods() -> None:
-    generators = np.random.randint(2, size=(6, 56), dtype=np.uint)
+    generators = np.random.randint(2, size=(6, 56), dtype=np.uint64)
     distance_default = qldpc.codes.distance.get_distance_classical(generators, block_size=3)
 
     # Tests should work with numpy < 2.0.0 so provide a backup `np.bitwise_count` implementation
@@ -412,8 +412,8 @@ def test_get_distance_classical_methods() -> None:
 
 
 def test_get_distance_quantum_methods() -> None:
-    stabilizers = np.random.randint(2, size=(4, 56), dtype=np.uint)
-    logical_ops = np.random.randint(2, size=(3, 56), dtype=np.uint)
+    stabilizers = np.random.randint(2, size=(4, 56), dtype=np.uint64)
+    logical_ops = np.random.randint(2, size=(3, 56), dtype=np.uint64)
     distance_default = qldpc.codes.distance.get_distance_quantum(
         logical_ops, stabilizers, block_size=3, homogeneous=True
     )
@@ -467,8 +467,8 @@ def test_get_distance_quantum_methods() -> None:
 
 
 def test_get_distance_quantum_methods_symplectic() -> None:
-    stabilizers = np.random.randint(2, size=(4, 56), dtype=np.uint)
-    logical_ops = np.random.randint(2, size=(3, 56), dtype=np.uint)
+    stabilizers = np.random.randint(2, size=(4, 56), dtype=np.uint64)
+    logical_ops = np.random.randint(2, size=(3, 56), dtype=np.uint64)
     distance_default = qldpc.codes.distance.get_distance_quantum(
         logical_ops, stabilizers, block_size=3, homogeneous=False
     )
